move work template menu item + wording
