+++ components/global_header/left_controls/product_menu/product_menu_list/index.ts
@@ -27,7 +27,6 @@ import {openModal} from 'actions/views/modals';
 import {ModalData} from 'types/actions';
 import {CloudProducts} from 'utils/constants';
 import {isCloudLicense} from 'utils/license_utils';
-import {areWorkTemplatesEnabled} from 'selectors/work_template';
 
 import ProductMenuList from './product_menu_list';
 
@@ -54,8 +53,6 @@ function mapStateToProps(state: GlobalState) {
     const step = getInt(state, OnboardingTaskCategory, OnboardingTasksName.VISIT_SYSTEM_CONSOLE, 0);
     const showVisitSystemConsoleTour = step === TaskNameMapToSteps[OnboardingTasksName.VISIT_SYSTEM_CONSOLE].STARTED;
     const enableCustomUserGroups = isCustomGroupsEnabled(state);
-
-    const showWorkTemplateButton = areWorkTemplatesEnabled(state);
 
     const subscription = getCloudSubscription(state);
     const license = getLicense(state);
@@ -92,7 +89,6 @@ function mapStateToProps(state: GlobalState) {
         enableCustomUserGroups,
         isStarterFree,
         isFreeTrial,
-        showWorkTemplateButton,
     };
 }
