+++ components/sidebar/add_channel_dropdown.tsx
@@ -11,6 +11,8 @@ import Menu from 'components/widgets/menu/menu';
 import OverlayTrigger from 'components/overlay_trigger';
 import Tooltip from 'components/tooltip';
 import {CreateAndJoinChannelsTour, InvitePeopleTour} from 'components/tours/onboarding_tour';
+import {ModalIdentifiers} from 'utils/constants';
+import WorkTemplateModal from 'components/work_templates';
 
 type Props = {
     canCreateChannel: boolean;
@@ -28,6 +30,7 @@ type Props = {
     isAddChannelOpen: boolean;
     openAddChannelOpen: (open: boolean) => void;
     canCreateCustomGroups: boolean;
+    showWorkTemplateButton: boolean;
 };
 
 const AddChannelDropdown = ({
@@ -45,6 +48,7 @@ const AddChannelDropdown = ({
     isAddChannelOpen,
     openAddChannelOpen,
     canCreateCustomGroups,
+    showWorkTemplateButton,
 }: Props) => {
     const intl = useIntl();
 
@@ -61,6 +65,21 @@ const AddChannelDropdown = ({
                 {showInviteTutorialTip && <InvitePeopleTour/>}
             </Menu.Group>
         );
+
+        let workTemplate;
+        if (showWorkTemplateButton) {
+            workTemplate = (
+                <Menu.ItemToggleModalRedux
+                    id='work-template'
+                    modalId={ModalIdentifiers.WORK_TEMPLATES}
+                    dialogType={WorkTemplateModal}
+                    text={intl.formatMessage({id: 'sidebar_left.add_channel_dropdown.work_template', defaultMessage: 'Create from a template'})}
+                    extraText={intl.formatMessage({id: 'sidebar_left.add_channel_dropdown.work_template_extra', defaultMessage: 'Set up a channel with linked boards, and playbooks'})}
+                    icon={<i className='icon-layers-outline'/>}
+                    className='work-template'
+                />
+            );
+        }
 
         let joinPublicChannel;
         if (canJoinPublicChannel) {
@@ -123,6 +142,7 @@ const AddChannelDropdown = ({
         return (
             <>
                 <Menu.Group>
+                    {workTemplate}
                     {joinPublicChannel}
                     {createChannel}
                     {createDirectMessage}
