+++ i18n/en.json
@@ -4184,7 +4184,6 @@
   "navbar_dropdown.userGroups.modal.titleEndUser": "User groups available in paid plans",
   "navbar_dropdown.userGroups.tooltip.cloudFreeTrial": "During your trial you are able to create user groups. These user groups will be archived after your trial.",
   "navbar_dropdown.viewMembers": "View Members",
-  "navbar_dropdown.workTemplate": "Work Templates",
   "navbar.addGroups": "Add Groups",
   "navbar.addMembers": "Add Members",
   "navbar.preferences": "Notification Preferences",
@@ -4766,6 +4765,8 @@
   "sidebar_left.add_channel_dropdown.dropdownAriaLabel": "Add Channel Dropdown",
   "sidebar_left.add_channel_dropdown.invitePeople": "Invite People",
   "sidebar_left.add_channel_dropdown.invitePeopleExtraText": "Add people to the team",
+  "sidebar_left.add_channel_dropdown.work_template": "Create from a template",
+  "sidebar_left.add_channel_dropdown.work_template_extra": "Set up a channel with linked boards, and playbooks",
   "sidebar_left.channel_filter.filterByUnread": "Filter by unread",
   "sidebar_left.channel_filter.filterUnreadAria": "unreads filter",
   "sidebar_left.channel_filter.showAllChannels": "Show all channels",
