+++ sass/layout/_sidebar-left.scss
@@ -285,13 +285,15 @@ $sidebarOpacityAnimationDuration: 0.15s;
         height: 30px;
 
         .dropdown-menu {
-            min-width: 0;
+            min-width: 250px;
             margin-top: 0;
             line-height: 19px;
 
             span {
                 color: rgba(var(--center-channel-color-rgb), 0.9);
-                white-space: nowrap;
+                &:not(.MenuItem__help-text) {
+                    white-space: nowrap;
+                }
             }
 
             button {
